Moved all bussiness logic to a separate provider file
